backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pyjwt==2.8.0
bcrypt==4.0.1
passlib==1.7.4
python-dotenv==1.0.0
websockets==12.0
pydantic==2.5.3
